+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.flutter_1">
+
+    <uses-permission android:name="android.permission.INTERNET"/>
    <application
         android:label="flutter_1"
         android:name="${applicationName}"
