+++ assets/file/catalog.json
@@ -30,7 +30,7 @@
       "desc": "Sony Playstation 5th generation",
       "price": 500,
       "color": "#544ee4",
-      "image": "https://i1.wp.com/freepngimages.com/wp-content/uploads/2020/07/Playstation-5-games-console-transparent-background-png-image.png?fit=1000%2C1000"
+      "image": "https://www.telstra.com.au/content/dam/tcom/lego/2021/entertainment/content-xbox-series-s-controller-white-656x370.jpg"
     },
     {
       "id": 5,
