+++ lib/pages/home_detail_page.dart
@@ -1,0 +1,71 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
+import 'package:flutter/material.dart';
+import 'package:flutter_1/widgets/themes.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+import '../models/catalog.dart';
+
+class HomeDetailPage extends StatelessWidget {
+  final Item catalog;
+  HomeDetailPage({
+    Key? key,
+    required this.catalog,
+  }) : assert(catalog != null);
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      bottomNavigationBar: Container(
+        color: Colors.white,
+        child: ButtonBar(
+          alignment: MainAxisAlignment.spaceBetween,
+          buttonPadding: EdgeInsets.zero,
+          children: [
+            "\$${catalog.price}".text.bold.xl4.red800.make(),
+            ElevatedButton(
+              onPressed: () {},
+              style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all(
+                    MyTheme.darkBluishColor,
+                  ),
+                  shape: MaterialStateProperty.all(StadiumBorder())),
+              child: "Buy".text.xl.make(),
+            ).wh(100, 30)
+          ],
+        ).p32(),
+      ),
+      appBar: AppBar(),
+      backgroundColor: MyTheme.creamcolor,
+      body: SafeArea(
+        bottom: false,
+        child: Column(
+          children: [
+            Hero(
+              tag: Key(catalog.id.toString()),
+              child: Image.network(catalog.image),
+            ).h32(context),
+            Expanded(
+                child: VxArc(
+              arcType: VxArcType.CONVEY,
+              edge: VxEdge.TOP,
+              height: 30.0,
+              child: Container(
+                width: context.screenWidth,
+                color: Colors.white,
+                child: Column(
+                  children: [
+                    catalog.name.text.xl4
+                        .color(MyTheme.darkBluishColor)
+                        .bold
+                        .make(),
+                    catalog.desc.text.textStyle(context.captionStyle).xl.make(),
+                    10.heightBox,
+                  ],
+                ).py64(),
+              ),
+            ))
+          ],
+        ),
+      ),
+    );
+  }
+}
